Add image histogram

+++ Image/Image.py
@@ -587,9 +587,9 @@ class ImageTensor(Tensor):
     def histo_equalization(self, in_place=False, filtering=False, clahe=False):
         out = in_place_fct(self, in_place)
         if filtering:
-            out.data = equalize(out)
+            out.data = equalize(out.to(torch.uint8))
         elif clahe:
-            out.data = equalize_clahe(out)
+            out.data = equalize_clahe(out.to(torch.uint8))
         else:
             hist = out.hist()
             mini, maxi = hist.clip()
